requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.0
streamlit==1.32.0
plotly==5.20.0
folium==0.16.0
streamlit-folium==0.19.0
pyarrow==15.0.0
python-dotenv==1.0.1
requests==2.31.0
joblib==1.3.2
pytest==8.0.0
scipy==1.12.0
matplotlib==3.8.0
seaborn==0.13.2
